Refactor Folder structure - UD
