+++ src/EPR.Calculator.API.UnitTests/Builder/CalcResultBuilderTests.cs
@@ -1,4 +1,5 @@
 using EPR.Calculator.API.Builder.LateReportingTonnages;
+using EPR.Calculator.API.Builder.OnePlusFourApportionment;
 using EPR.Calculator.API.CommsCost;
 
 namespace EPR.Calculator.API.UnitTests.Builder
