+++ src/EPR.Calculator.API.UnitTests/Builder/CalcResultOnePlusFourApportionmentBuilderTest.cs
@@ -1,5 +1,4 @@
-﻿using EPR.Calculator.API.Builder;
-using EPR.Calculator.API.Constants;
+﻿using EPR.Calculator.API.Constants;
 using EPR.Calculator.API.Data;
 using EPR.Calculator.API.Dtos;
 using EPR.Calculator.API.Models;
@@ -12,6 +11,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using EPR.Calculator.API.Builder.OnePlusFourApportionment;
 
 namespace EPR.Calculator.API.UnitTests.Builder
 {
