+++ src/EPR.Calculator.API/Builder/CalcResultBuilder.cs
@@ -1,12 +1,11 @@
 ﻿using EPR.Calculator.API.Builder.LaDisposalCost;
 using EPR.Calculator.API.Builder.Lapcap;
-using EPR.Calculator.API.CommsCost;
 using EPR.Calculator.API.Builder.LateReportingTonnages;
+using EPR.Calculator.API.Builder.OnePlusFourApportionment;
 using EPR.Calculator.API.Builder.ParametersOther;
-using EPR.Calculator.API.Data;
+using EPR.Calculator.API.CommsCost;
 using EPR.Calculator.API.Dtos;
 using EPR.Calculator.API.Models;
-using Microsoft.VisualStudio.TestPlatform.ObjectModel;
 
 namespace EPR.Calculator.API.Builder
 {
