+++ src/EPR.Calculator.API/Builder/OnePlusFourApportionment/ICalcResultOnePlusFourApportionmentBuilder.cs
@@ -1,0 +1,9 @@
+﻿using EPR.Calculator.API.Dtos;
+using EPR.Calculator.API.Models;
+
+namespace EPR.Calculator.API.Builder.OnePlusFourApportionment;
+
+public interface ICalcResultOnePlusFourApportionmentBuilder
+{
+    public CalcResultOnePlusFourApportionment Construct(CalcResultsRequestDto resultsRequestDto, CalcResult calcResult);
+}
