+++ src/EPR.Calculator.API/Program.cs
@@ -5,6 +5,7 @@ using EPR.Calculator.API.Builder.LaDisposalCost;
 using EPR.Calculator.API.Builder.Lapcap;
 using EPR.Calculator.API.Builder.ParametersOther;
 using EPR.Calculator.API.Builder.LateReportingTonnages;
+using EPR.Calculator.API.Builder.OnePlusFourApportionment;
 using EPR.Calculator.API.CommsCost;
 using EPR.Calculator.API.Constants;
 using EPR.Calculator.API.Data;
